# Refactor vectorization and preloading for pointwise fusions