--- a/src/targets/gpu/jit/pointwise.cpp
+++ b/src/targets/gpu/jit/pointwise.cpp
@@ -6,6 +6,7 @@
 #include <migraphx/cpp_generator.hpp>
 #include <migraphx/ranges.hpp>
 #include <migraphx/reduce_dims.hpp>
+#include <migraphx/permutation.hpp>
 #include <migraphx/stringutils.hpp>
 #include <migraphx/dead_code_elimination.hpp>
 #include <migraphx/eliminate_common_subexpression.hpp>
@@ -29,7 +30,7 @@ extern "C" {
 __global__ void kernel(${params}) 
 {
     auto idx = make_index();
-    pointwise(idx, auto_preload(idx), vectorize<${vec_size}, ${vec_axis}>())(${lambda}, ${args});
+    pointwise(idx, auto_preload<${preloads}>(idx), vectorize<${vec_size}, ${axis}>())(${lambda}, ${args});
 }
     
 }
@@ -42,12 +43,46 @@ struct pointwise_compiler : compiler<pointwise_compiler>
 {
     std::vector<std::string> names() const { return {"pointwise"}; }
 
-    static std::size_t oversubscribe(const std::vector<shape>& inputs)
+    static std::size_t oversubscribe(bool preloading)
     {
-        if(std::any_of(inputs.begin(), inputs.end(), [](const auto& s) { return s.broadcasted(); }))
+        if (preloading)
             return 1;
         else
             return 256;
+    }
+    static std::size_t find_fast_axis(const std::vector<shape>& inputs)
+    {
+        auto permutation = find_permutation(inputs);
+        auto it = std::max_element(permutation.begin(), permutation.end());
+        return it - permutation.begin();
+    }
+    static std::vector<bool> preload(std::size_t axis, const std::vector<shape>& inputs)
+    {
+        const std::size_t max_lds_bytes = 4096;
+        std::vector<bool> result;
+        std::transform(inputs.begin(), inputs.end(), std::back_inserter(result), [&](const shape& input) {
+            return input.strides()[axis] == 0;
+        });
+        auto bytes = std::inner_product(inputs.begin(), inputs.end(), result.begin(), std::size_t{0}, std::plus<>{}, [](const shape& s, bool b) -> std::size_t {
+            if (b)
+                return s.bytes();
+            return 0;
+        });
+        if (bytes < max_lds_bytes)
+            return result;
+        // TODO: Try to partially preload items
+        std::fill(result.begin(), result.end(), false);
+        return result;    
+    }
+    static std::string preload_str(const std::vector<bool>& bs)
+    {
+        std::vector<std::string> bool_strs;
+        std::transform(bs.begin(), std::prev(bs.end()), std::back_inserter(bool_strs), [](bool b) {
+            if (b)
+                return "true";
+            return "false";
+        });
+        return "false, " + join_strings(bool_strs, ", ");
     }
     static std::vector<std::size_t> vector_sizes(const std::vector<shape>& inputs)
     {
@@ -58,21 +93,8 @@ struct pointwise_compiler : compiler<pointwise_compiler>
             return {2};
         return {4, 2};
     }
-    static auto vectorize_elements(const std::vector<shape>& inputs)
+    static auto vectorize_elements(std::size_t axis, const std::vector<shape>& inputs)
     {
-        std::size_t n    = 1;
-        std::size_t rank = inputs.front().lens().size();
-        std::size_t axis = rank;
-        for(const auto& input : inputs)
-        {
-            if(input.broadcasted())
-                continue;
-            axis = std::distance(input.strides().begin(),
-                                 std::find(input.strides().begin(), input.strides().end(), 1));
-            break;
-        }
-        if(axis >= rank)
-            return std::make_pair(n, axis);
         auto sizes = vector_sizes(inputs);
         std::vector<std::size_t> max_vec_size;
         std::transform(inputs.begin(),
@@ -89,8 +111,7 @@ struct pointwise_compiler : compiler<pointwise_compiler>
                                return *it;
                            return 1;
                        });
-        n = *std::min_element(max_vec_size.begin(), max_vec_size.end());
-        return std::make_pair(n, axis);
+        return *std::min_element(max_vec_size.begin(), max_vec_size.end());
     }
     operation compile_op(context& ctx, const std::vector<shape>& inputs, const value& v) const
     {
@@ -99,16 +120,20 @@ struct pointwise_compiler : compiler<pointwise_compiler>
         options.output            = inputs.back();
         options.virtual_inputs    = reduce_dims(inputs);
         options.params            = "-Wno-float-equal";
-        auto [vec_size, vec_axis] = vectorize_elements(options.virtual_inputs);
+        auto axis = find_fast_axis(options.virtual_inputs);
+        auto vec_size = vectorize_elements(axis, options.virtual_inputs);
+        auto preloads = preload(axis, options.virtual_inputs);
+        auto is_preloading = std::accumulate(preloads.begin(), preloads.end(), false, std::logical_or<>{});
         options.set_launch_params(
             v,
-            compute_global_for(ctx, options.output.elements() / vec_size, oversubscribe(inputs)));
+            compute_global_for(ctx, options.output.elements() / vec_size, oversubscribe(is_preloading)));
         auto src = interpolate_string(pointwise_kernel,
                                       {{"params", enum_params(inputs.size(), "void * private_p")},
                                        {"args", enum_params(inputs.size(), "private_p")},
                                        {"lambda", v.at("lambda").to<std::string>()},
                                        {"vec_size", std::to_string(vec_size)},
-                                       {"vec_axis", std::to_string(vec_axis)},
+                                       {"axis", std::to_string(axis)},
+                                       {"preloads", preload_str(preloads)},
                                        {"preamble", v.get("preamble", std::string{})}});
         return compile_hip_code_object(src, options);
     }