… src/targets/gpu/kernels/include/migraphx/kernels/preload.hpp
… src/targets/gpu/kernels/include/migraphx/kernels/vec.hpp
… src/targets/gpu/kernels/include/migraphx/kernels/vectorize.hpp