This improves performance. For add_gelu in bert its 4x faster and for mul_add its 50% faster than what is on develop:

```
// This PR
gpu::code_object[code_object=13736,symbol_name=kernel,global=14155776,local=1024,]: 0.243313ms
// Develop
gpu::code_object[code_object=13736,symbol_name=kernel,global=153600,local=1024,]: 0.380066ms
gpu::mul_add: 0.475723ms


// This PR
gpu::code_object[code_object=14176,symbol_name=kernel,global=9437184,local=1024,]: 0.201901ms
// Develop
gpu::code_object[code_object=14016,symbol_name=kernel,global=153600,local=1024,]: 0.89556ms
gpu::add_gelu: 3.04435ms
```

This matches the performance on branch_for_ort2 for these same operators.