@causten 2022-04-25
This change does improve performance.  Closer to what is seen in branch_for_ort2.